# Pluggable routers + mmh3 faster hashed router