--- a/lib/carbon/tests/relay-rules.conf
+++ b/lib/carbon/tests/relay-rules.conf
@@ -1,0 +1,3 @@
+[default]
+destinations = foo:124:a
+default = true