This merge request adds:
* Mechanisms to allow pluggable routers
* A new set of mmh3 based faster routers

On a heavily loaded setup using consistent-hashing the router ends up taking a good chunk of the CPU time, and the current router takes more time as the number of nodes in the ring grows. It's not super important to reduce the number of rebalances when the ring is changed because carbonate complexity depends on the total number of metric.

Here is the output of the micro-benchmark embedded in this pull request
```
 consistent-hashing  deplication_factor: 1 diverse_replicas: 1 n_destinations: 1     datapoints: 100000 msecs: 246
 consistent-hashing  deplication_factor: 1 diverse_replicas: 1 n_destinations: 16    datapoints: 100000 msecs: 512
 consistent-hashing  deplication_factor: 1 diverse_replicas: 1 n_destinations: 48    datapoints: 100000 msecs: 526
 consistent-hashing  deplication_factor: 1 diverse_replicas: 0 n_destinations: 1     datapoints: 100000 msecs: 226
 consistent-hashing  deplication_factor: 1 diverse_replicas: 0 n_destinations: 16    datapoints: 100000 msecs: 562
 consistent-hashing  deplication_factor: 1 diverse_replicas: 0 n_destinations: 48    datapoints: 100000 msecs: 573
 consistent-hashing  deplication_factor: 4 diverse_replicas: 1 n_destinations: 1     datapoints: 100000 msecs: 229
 consistent-hashing  deplication_factor: 4 diverse_replicas: 1 n_destinations: 16    datapoints: 100000 msecs: 783
 consistent-hashing  deplication_factor: 4 diverse_replicas: 1 n_destinations: 48    datapoints: 100000 msecs: 803
 consistent-hashing  deplication_factor: 4 diverse_replicas: 0 n_destinations: 1     datapoints: 100000 msecs: 221
 consistent-hashing  deplication_factor: 4 diverse_replicas: 0 n_destinations: 16    datapoints: 100000 msecs: 931
 consistent-hashing  deplication_factor: 4 diverse_replicas: 0 n_destinations: 48    datapoints: 100000 msecs: 818
 fast-aggregated-hashing  deplication_factor: 1 diverse_replicas: 1 n_destinations: 1     datapoints: 100000 msecs: 429
 fast-aggregated-hashing  deplication_factor: 1 diverse_replicas: 1 n_destinations: 16    datapoints: 100000 msecs: 412
 fast-aggregated-hashing  deplication_factor: 1 diverse_replicas: 1 n_destinations: 48    datapoints: 100000 msecs: 408
 fast-aggregated-hashing  deplication_factor: 1 diverse_replicas: 0 n_destinations: 1     datapoints: 100000 msecs: 371
 fast-aggregated-hashing  deplication_factor: 1 diverse_replicas: 0 n_destinations: 16    datapoints: 100000 msecs: 430
 fast-aggregated-hashing  deplication_factor: 1 diverse_replicas: 0 n_destinations: 48    datapoints: 100000 msecs: 426
 fast-aggregated-hashing  deplication_factor: 4 diverse_replicas: 1 n_destinations: 1     datapoints: 100000 msecs: 380
 fast-aggregated-hashing  deplication_factor: 4 diverse_replicas: 1 n_destinations: 16    datapoints: 100000 msecs: 621
 fast-aggregated-hashing  deplication_factor: 4 diverse_replicas: 1 n_destinations: 48    datapoints: 100000 msecs: 618
 fast-aggregated-hashing  deplication_factor: 4 diverse_replicas: 0 n_destinations: 1     datapoints: 100000 msecs: 373
 fast-aggregated-hashing  deplication_factor: 4 diverse_replicas: 0 n_destinations: 16    datapoints: 100000 msecs: 605
 fast-aggregated-hashing  deplication_factor: 4 diverse_replicas: 0 n_destinations: 48    datapoints: 100000 msecs: 607
 fast-hashing  deplication_factor: 1 diverse_replicas: 1 n_destinations: 1     datapoints: 100000 msecs: 306
 fast-hashing  deplication_factor: 1 diverse_replicas: 1 n_destinations: 16    datapoints: 100000 msecs: 310
 fast-hashing  deplication_factor: 1 diverse_replicas: 1 n_destinations: 48    datapoints: 100000 msecs: 309
 fast-hashing  deplication_factor: 1 diverse_replicas: 0 n_destinations: 1     datapoints: 100000 msecs: 293
 fast-hashing  deplication_factor: 1 diverse_replicas: 0 n_destinations: 16    datapoints: 100000 msecs: 366
 fast-hashing  deplication_factor: 1 diverse_replicas: 0 n_destinations: 48    datapoints: 100000 msecs: 354
 fast-hashing  deplication_factor: 4 diverse_replicas: 1 n_destinations: 1     datapoints: 100000 msecs: 311
 fast-hashing  deplication_factor: 4 diverse_replicas: 1 n_destinations: 16    datapoints: 100000 msecs: 525
 fast-hashing  deplication_factor: 4 diverse_replicas: 1 n_destinations: 48    datapoints: 100000 msecs: 530
 fast-hashing  deplication_factor: 4 diverse_replicas: 0 n_destinations: 1     datapoints: 100000 msecs: 297
 fast-hashing  deplication_factor: 4 diverse_replicas: 0 n_destinations: 16    datapoints: 100000 msecs: 470
 fast-hashing  deplication_factor: 4 diverse_replicas: 0 n_destinations: 48    datapoints: 100000 msecs: 495
 aggregated-consistent-hashing  deplication_factor: 1 diverse_replicas: 1 n_destinations: 1     datapoints: 100000 msecs: 349
 aggregated-consistent-hashing  deplication_factor: 1 diverse_replicas: 1 n_destinations: 16    datapoints: 100000 msecs: 716
 aggregated-consistent-hashing  deplication_factor: 1 diverse_replicas: 1 n_destinations: 48    datapoints: 100000 msecs: 755
 aggregated-consistent-hashing  deplication_factor: 1 diverse_replicas: 0 n_destinations: 1     datapoints: 100000 msecs: 339
 aggregated-consistent-hashing  deplication_factor: 1 diverse_replicas: 0 n_destinations: 16    datapoints: 100000 msecs: 761
 aggregated-consistent-hashing  deplication_factor: 1 diverse_replicas: 0 n_destinations: 48    datapoints: 100000 msecs: 731
 aggregated-consistent-hashing  deplication_factor: 4 diverse_replicas: 1 n_destinations: 1     datapoints: 100000 msecs: 339
 aggregated-consistent-hashing  deplication_factor: 4 diverse_replicas: 1 n_destinations: 16    datapoints: 100000 msecs: 989
 aggregated-consistent-hashing  deplication_factor: 4 diverse_replicas: 1 n_destinations: 48    datapoints: 100000 secs: 1.04423
 aggregated-consistent-hashing  deplication_factor: 4 diverse_replicas: 0 n_destinations: 1     datapoints: 100000 msecs: 339
 aggregated-consistent-hashing  deplication_factor: 4 diverse_replicas: 0 n_destinations: 16    datapoints: 100000 secs: 1.01963
 aggregated-consistent-hashing  deplication_factor: 4 diverse_replicas: 0 n_destinations: 48    datapoints: 100000 secs: 1.05064
```

![chart](https://cloud.githubusercontent.com/assets/1032963/19068104/644de804-8a21-11e6-8439-a7b66258a618.png)